Добавляет трип контроллер
